@@ -4,29 +4,30 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import com.android36kr.app.module.common.log.KrLog
 import com.google.firebase.analytics.FirebaseAnalytics
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.item_widget.view.*
+import com.qwlyz.androidstudy.databinding.ActivityMainBinding
+import com.qwlyz.androidstudy.widget.MessageEvent
+import com.yuwq.libs_common.viewBinding
 
 
 class MainActivity : AppCompatActivity() {
 
+    private val binding by viewBinding(ActivityMainBinding::inflate)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        val instance = FirebaseAnalytics.getInstance(this)
-
-        recycler_view.addItemDecoration(DividerItemDecoration(this, DividerItemDecoration.VERTICAL))
-        recycler_view.adapter = WidgetAdapter(instance)
+        setContentView(binding.root)
+        binding.recyclerView.addItemDecoration(DividerItemDecoration(this, DividerItemDecoration.VERTICAL))
+        binding.recyclerView.adapter = WidgetAdapter()
     }
 
-    class WidgetAdapter(val instance: FirebaseAnalytics) : RecyclerView.Adapter<WidgetViewHolder>() {
-        val dataList = PageWidget.values()
+    class WidgetAdapter : RecyclerView.Adapter<WidgetViewHolder>() {
+        private val dataList = PageWidget.values()
 
         override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WidgetViewHolder {
             return WidgetViewHolder(
@@ -37,9 +38,10 @@ class MainActivity : AppCompatActivity() {
         override fun getItemCount(): Int = dataList.size
 
         override fun onBindViewHolder(holder: WidgetViewHolder, position: Int) {
-            holder.itemView.text_view.text = dataList[position].title
+            holder.itemView.findViewById<TextView>(R.id.text_view).text = dataList[position].title
             holder.itemView.setOnClickListener {
                 WidgetActivity.start(holder.itemView.context, dataList[position])
+                MessageEvent.report()
             }
         }
 
